+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 # CHANGELOG
 
+<b>Version 2.2.10</b> - November 2023<br/>
+<p>
+- NEW: Hue Light: you can now enable the input/output PINs and send/receive commands to/from the light, via the msg flow, like msg.on={"on":true}. The option is "Node Input/Output PINs".<br/>
+- NEW: Hue Scene: you can now enable the input/output PINs and send/receive commands to/from the light, via the msg flow. The option is "Node Input/Output PINs".<br/>
+</p>
 <b>Version 2.2.9</b> - November 2023<br/>
 <p>
 - Fixed errors in Iobroker.<br/>